+++ src/designs/basic/screens/BasicHeader.tsx
@@ -1,5 +1,5 @@
 import { Dialog, DialogTrigger } from "@ui/dialog";
-import { Tooltip, TooltipContent, TooltipTrigger } from "@ui/tooltip";
+import { Tooltippy } from "@ui/tooltip";
 import { RotateCcw, SettingsIcon } from "lucide-react";
 import { ClearSessionDialogContent, SettingsDialogContent } from "./DialogContents";
 import { useLocale } from "@lib/index";
@@ -10,16 +10,13 @@ function SettingsDialog() {
     const locale = useLocale();
     return (
         <Dialog>
-            <Tooltip>
-                <TooltipContent side="bottom">
-                    {locale.get("settings")}
-                </TooltipContent>
-                <TooltipTrigger asChild>
-                    <DialogTrigger className="p-1.5 rounded-xl text-secondary-foreground flex-shrink-0">
-                        <SettingsIcon className="size-5" />
-                    </DialogTrigger>
-                </TooltipTrigger>
-            </Tooltip>
+            <Tooltippy
+                content={locale.get("settings")}
+            >
+                <DialogTrigger className="p-1.5 rounded-xl text-secondary-foreground flex-shrink-0">
+                    <SettingsIcon className="size-5" />
+                </DialogTrigger>
+            </Tooltippy>
             <SettingsDialogContent />
         </Dialog>
     );
@@ -31,16 +28,11 @@ function ResetConversationDialog() {
         <Dialog>
             {({ setOpen }) => (
                 <>
-                    <Tooltip>
-                        <TooltipContent side="bottom">
-                            {locale.get("reset-conversation")}
-                        </TooltipContent>
-                        <TooltipTrigger asChild>
-                            <DialogTrigger className="p-1 rounded-full text-secondary-foreground hover:brightness-110 flex-shrink-0">
-                                <RotateCcw className="size-4" />
-                            </DialogTrigger>
-                        </TooltipTrigger>
-                    </Tooltip>
+                    <Tooltippy content={locale.get("reset-conversation")}>
+                        <DialogTrigger className="p-1 rounded-full text-secondary-foreground hover:brightness-110 flex-shrink-0">
+                            <RotateCcw className="size-4" />
+                        </DialogTrigger>
+                    </Tooltippy>
                     <ClearSessionDialogContent setOpen={setOpen} />
                 </>
             )}